Fixed Parser::getParentDirectory not working properly on paths without a parent directory. WIll return an empty string in that case now.
Implemented FileLister::setCheckFolderDepth (functionality was there, the setter was just left empty)
Added FileLister::getSelectedType to return the type of the selected element (if it is a dir or not). This also fixes a crash when calling goUp before calling update or render (before a display has been created).
Added functions to modify the output of FileLister (enable/disable working path label, current dir item "." and parent dir item "..")